chore: icon prop
