+++ .changeset/popular-dodos-drive.md
@@ -1,0 +1,5 @@
+---
+"@shoplflow/base": patch
+---
+
+Icon HTMLAttributes 추가
