+++ packages/base/src/components/Icon/Icon.mdx
@@ -8,9 +8,11 @@ import * as IconStories from './Icon.stories';
 
 Icon Component를 레핑하는 컴포넌트입니다. 각 사이즈에 따라 아이콘 크기를 변경할 수 있습니다.
 
-## dangerouslySetInnerHTML
+### dangerouslySetInnerHTML
 
-`dangerouslySetInnerHTML`을 사용하여 아이콘을 렌더링할 수 있습니다.
+`dangerouslySetInnerHTML`을 사용하여 스토리북에서 아이콘을 미리 확인할 수 있습니다.
+
+> ‼️ 실제 코드에서는 dangerouslySetInnerHTML prop과 동일하게 동작합니다.
 
 svg를 string 타입으로 받아 렌더링합니다. `<svg>` 태그 **내부 요소**만 전달해야 합니다.
