+++ packages/base/src/components/Icon/Icon.types.ts
@@ -1,4 +1,5 @@
 import type { $Values } from '@shoplflow/utils';
+import type { HTMLAttributes } from 'react';
 import type { ColorTokenProps, IconSourceProps, SizeVariantProps } from '../../utils/type/ComponentProps';
 
 export const IconSizeVariants = {
@@ -12,9 +13,11 @@ export const IconSizeVariants = {
 export type IconSizeVariantsType = $Values<typeof IconSizeVariants>;
 
 export interface IconProps extends IconOptionProps {}
-export interface IconOptionProps extends SizeVariantProps<IconSizeVariantsType>, IconSourceProps, ColorTokenProps {
-  dangerouslySetInnerHTML?: string;
-}
+export interface IconOptionProps
+  extends SizeVariantProps<IconSizeVariantsType>,
+    IconSourceProps,
+    ColorTokenProps,
+    Omit<HTMLAttributes<SVGSVGElement>, 'color'> {}
 
 export type DangerouslySetInnerHTML =
   | (string & {
